HMIS.cpp
Hospital Management System Project in C++ Using Classes was create using a simple console application. This Program is open to all hospital users without the need for login information in this program. The system user can select the category in which they specialize by simply entering a series of numeric keys.